requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
# -i https://pypi.douban.com/simple/ 

fabric3
